Create a documentation procedure for repo
As Sphinx is used to autogenerate documentation, an automated workflow
should be developed for devs to create the docs in a standardised manner.

For this reason, the setup/venv_setup.sh bash script was updated to add
install Calibre, which is used to convert .epub files, that Sphinx can
generate, to .pdf.

The docs/automake.sh file was created to automate the generation process
of the HTML and PDF files.

Finally, the README.md was updated to include instructions on generating
documentation automatically.