Make OutputManager a non-Singleton.
The shared state of the singleton was causing test_prospector in test_biru_prospector.py to fail, because the files_to_ignore field was being populated by another test with irrelevant ignore-list entries. The test would pass when run in isolation, but not when the tests were run in batch.