IngestionFactory cleanup
Make config a mandatory argument, instead of trying to read environment variables for it, in order to reduce the number of places where we interact with environment variables (which are effectively globals).

No longer exit the entire program (!) if the config is not passed, as this is not a decision for a low-level component to make.

Make IngestionFactory a non-Singleton, to fix some static analysis issues elsewhere.